Unfinished announcement group creation (migrations required)

+++ src/Controller/AdminController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Announcement;
 use App\Entity\Group;
+use App\Form\AnnouncementForm;
 use App\Form\GroupForm;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,12 +15,27 @@ class AdminController extends AbstractController {
 		$this->denyAccessUnlessGranted('ROLE_ADMIN');
 		$em = $this->getDoctrine()->getManager();
 		
+		//creating the new announcement.
+		$announcement = new Announcement();
+		$form = $this->createForm(AnnouncementForm::class, $announcement);
+		$form->handleRequest($request);
+		if($form->isSubmitted() && $form->isValid()) {
+			$announcement = $form->getData();
+			$announcement->setUser($this->getUser());
+			try {
+				$announcement->setCreationDate(new \DateTime());
+			} catch(\Exception $e) {
+				throw $e;
+			}
+			$em->persist($announcement);
+			$em->flush();
+		}
 		$blocks = $em->getRepository(Announcement::class)->findAll();
 		
-		return $this->render('admin/block.html.twig', ['blocks' => $blocks]);
+		return $this->render('admin/block.html.twig', ['blocks' => $blocks, 'form' => $form->createView()]);
 	}
 	
-	public function group(Request $request) {
+	public function group(Request $request, $message) {
 		$this->denyAccessUnlessGranted('ROLE_ADMIN');
 		$em = $this->getDoctrine()->getManager();
 		
@@ -31,7 +47,6 @@ class AdminController extends AbstractController {
 			$group = $form->getData();
 			$em->persist($group);
 			$em->flush();
-			
 		}
 		
 		$groups = $em->getRepository(Group::class)->findAll();
@@ -39,18 +54,23 @@ class AdminController extends AbstractController {
 		return $this->render('admin/group.html.twig', ['groups' => $groups, 'form' => $form->createView()]);
 	}
 	
-	public function group_delete($group) {
+	public function group_delete($group_id) {
 		$this->denyAccessUnlessGranted('ROLE_ADMIN');
 		
+		$em = $this->getDoctrine()->getManager();
+		
+		$repository = $this->getDoctrine()->getRepository(Group::class);
+		$group = $repository->find($group_id);
+		if (!$group) {
+			throw $this->createNotFoundException('No product found for id ' . $group_id);
+		}
 		if(count($group->getChildren()) > 0) {
 			return $this->render('admin/group_delete.html.twig', ['message' => "Group" . $group->getName() . " can't be removed because it contains children."]);
 		}
-		
-		$em = $this->getDoctrine()->getManager();
 		$em->remove($group);
 		$em->flush();
 		
-		return $this->render('admin/group_delete.html.twig', ['message' => "Group" . $group->getName() . " Removed."]);
+		return $this->redirectToRoute("admin_group", array('message' => "Group" . $group->getName() . " can't be removed because it contains children."));
 		
 	}
 }
